sidecar/requirements.txt
mem0ai[local]>=2.0.0,<3.0.0
fastapi>=0.136.0,<1.0.0
uvicorn>=0.46.0,<1.0.0
sentence-transformers>=5.0.0,<6.0.0
boto3>=1.42.0,<2.0.0
